Adding a more generic behavior for controls, use it in the file browser.
