+++ GoogleCloudExtension/GoogleCloudExtension.Utils/GoogleCloudExtension.Utils.csproj
@@ -57,6 +57,7 @@
     <Compile Include="AsyncPropertyValue.cs" />
     <Compile Include="BindableList.cs" />
     <Compile Include="BooleanConverter.cs" />
+    <Compile Include="ControlBehaviors.cs" />
     <Compile Include="DataGridBehaviors.cs" />
     <Compile Include="DataGridColumnCustomSorter.cs" />
     <Compile Include="Disposable.cs" />
